Add files via upload
This is a simple online Pharmacy management website where patients or doctors can buy medicines and pharmaceutical companies can add or delete products, view orders of customers etc. It was done as an university project. There are some limitations like an option was supposed to be added where patients can click on the pharmaceutical company and view their products. But now, patients can only see the list of the companies.